Update FlowchartPlugin.pm
I noticed, that changing the textsize in a macro, will only change the line distance, but the size of the letters remains the same (17pt).
This is caused, because the $textSite was insterted into $styleText too early in the program.
Between that action and the use of $styleText, the variable $textSize may be changed by parametern in the macro. But that change doesn't apply to $styleText.
My change is, setting $styleText right before the line, where it's used.
After that, the plugin works fime on my installation.